BOOTLOADER WORKS! detects and runs user code!
Bootloader is entered on powerup or reset if it does not find code at
address 0x1000 (will change to 0x0200 when I remove logging) or if pin
RA3 is grounded. (MCLRE config flag set to OFF)

"Code present" == "the lower 8 bits of the first word are not 0xFF".
Requires fewer instructions than fetching the entire first word and
checking if it's 0x3FFF.